[UPDATE] refactor RevisionOption and WrongExpectedVersion Error.
